Reorganizing code for Zuul printer
